requirements.txt
torch>=2.9.0
torchvision>=0.15.0
torchaudio>=2.2.0
rembg[gpu]
xatlas
trimesh
numpy
scipy
pillow
gradio
requests
pytest
